+++ ceng/__init__.py
@@ -1,3 +1,3 @@
 """Tool for Coding Tasks Specific to Civil Engineering"""
 
-__version__ = "0.3.1"
+__version__ = "0.4"